# Fase-I/guidelines/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sistema de Análisis de Sentimiento de Reviews – Hotel OLH",
    "description": "API REST para el pipeline de NLP y el dashboard ejecutivo de análisis de sentimiento de reseñas hoteleras. Cubre los servicios: Data Verification & Cleaner, Sentiment Prediction, ABSAService (ABSA sobre tópicos hoteleros → `review_topicos`), Topic Identification (alternativa al ABSA para la misma etapa `topics_identified`), Metric Calculation (`POST /metricas/calcular` y `POST /metricas/recalcular` en MetricProcessor), Dashboard Backend (lecturas `GET /metricas/*`, `POST /metricas/recalcular-semestre` que orquesta recálculos vía MetricProcessor, `GET /metricas/archivos/entrada`, `POST /metricas/archivos/upload` y Reviews Explorer) y el listado filtrado de reviews.",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Tesis OLH",
      "email": "admin@olh.com"
    }
  },
  "servers": [
    {
      "url": "https://api.olh.com",
      "description": "Producción"
    },
    {
      "url": "http://localhost:8000",
      "description": "Desarrollo local (genérico)"
    },
    {
      "url": "http://localhost:5002",
      "description": "Dashboard Backend (local): GET /api/v1/metricas/*, POST /api/v1/metricas/recalcular-semestre, GET /api/v1/metricas/archivos/entrada y POST /api/v1/metricas/archivos/upload"
    },
    {
      "url": "http://localhost:5004",
...
    },
    {
      "url": "http://localhost:5003",
...
    },
    {
...
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Iniciar sesión",
        "description": "Autentica usuario activo y retorna `access_token` + `refresh_token`.",
        "operationId": "authLogin",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Renovar access token",
        "description": "Requiere `Authorization: Bearer {refresh_token}` y retorna un nuevo access token.",
        "operationId": "authRefresh",
        "responses": {
          "200": {
            "description": "Token renovado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Cerrar sesión",
        "description": "Revoca access token actual y refresh token enviado en body.",
        "operationId": "authLogout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refresh_token"
                ],
                "properties": {
                  "refresh_token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens revocados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Token ausente o inválido",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EnvelopeError"
            },
            "example": {
              "ok": false,
              "data": null,
              "error": {
                "codigo": "NO_AUTORIZADO",
                "mensaje": "Token de autenticación inválido o expirado"
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Error interno del servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EnvelopeError"
            },
            "example": {
              "ok": false,
              "data": null,
              "error": {
                "codigo": "ERROR_INTERNO",
                "mensaje": "Error inesperado en el procesamiento"
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "required": [
          "ok",
          "data",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "nullable": true
          },
          "error": {
            "type": "object",
            "nullable": true,
            "example": null
          }
        }
      },
      "EnvelopeError": {
        "type": "object",
        "required": [
          "ok",
          "data",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "object",
            "nullable": true,
            "example": null
          },
          "error": {
            "type": "object",
            "required": [
              "codigo",
              "mensaje"
            ],
            "properties": {
              "codigo": {
                "type": "string",
                "example": "ARCHIVO_NO_ENCONTRADO"
              },
              "mensaje": {
                "type": "string",
                "example": "No se encontró el recurso solicitado"
              }
            }
          }
        }
      },
...
    }
  }
}